# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Foundry Academy LMS API
  version: 1.0.0
  description: |
    Canonical API contract for the Foundry Academy LMS.

    Authentication uses the HTTP-only `token` cookie. In Postman, run Login or
    Verify email OTP first and allow Postman's cookie jar to retain the cookie.
    Admin and super-admin requests require an account with the matching backend
    permission.
servers:
  - url: "{baseUrl}"
    description: Selected Foundry LMS API server
    variables:
      baseUrl:
        default: http://localhost:5000
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/categories:
    get:
      tags: [04 Categories]
      operationId: listCategoriesAdmin
      summary: List categories for administration
      x-access: ADMIN
      parameters:
        - $ref: "#/components/parameters/ServiceIdFilter"
        - $ref: "#/components/parameters/CatalogStatusFilter"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          $ref: "#/components/responses/Success"
    post:
      tags: [04 Categories]
      operationId: createCategory
      summary: Create category draft
      x-access: ADMIN
      requestBody:
        $ref: "#/components/requestBodies/CreateCategory"
      responses:
        "201":
          $ref: "#/components/responses/Created"
  /api/v1/categories/{categoryId}:
    get:
      tags: [04 Categories]
      operationId: getCategoryAdmin
      summary: Get category for administration
      x-access: ADMIN
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          $ref: "#/components/responses/Success"
    patch:
      tags: [04 Categories]
      operationId: updateCategory
      summary: Update category
      x-access: ADMIN
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      requestBody:
        $ref: "#/components/requestBodies/UpdateCategory"
      responses:
        "200":
          $ref: "#/components/responses/Success"
    delete:
      tags: [04 Categories]
      operationId: deleteCategoryPermanently
      summary: "[DESTRUCTIVE] Permanently delete archived category"
      x-access: SUPER_ADMIN
      description: Super-admin only. Deletes only archived catalog setup without course curriculum or learner history.
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/CatalogDeletionBlocked"
  /api/v1/categories/{categoryId}/deletion-impact:
    get:
      tags: [04 Categories]
      operationId: getCategoryDeletionImpact
      summary: Preview whether permanent category deletion is safe
      x-access: SUPER_ADMIN
      description: Super-admin only. Aggregates child course-intake curriculum and protected learner history without changing data.
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /api/v1/categories/{categoryId}/publish:
    patch:
      tags: [04 Categories]
      operationId: publishCategory
      summary: Publish category
      x-access: SUPER_ADMIN
      description: Super-admin only.
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /api/v1/categories/{categoryId}/unpublish:
    patch:
      tags: [04 Categories]
      operationId: unpublishCategory
      summary: Unpublish category
      x-access: SUPER_ADMIN
      description: Super-admin only.
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /api/v1/categories/{categoryId}/archive:
    patch:
      tags: [04 Categories]
      operationId: archiveCategory
      summary: Archive category and child courses
      x-access: SUPER_ADMIN
      description: Super-admin only. Preserves existing learner records and session content.
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /api/v1/categories/{categoryId}/unarchive:
    patch:
      tags: [04 Categories]
      operationId: unarchiveCategory
      summary: Restore category as draft
      x-access: SUPER_ADMIN
      description: Super-admin only. Child courses remain archived.
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          $ref: "#/components/responses/Success"
...
  /api/v1/courses:
    get:
      tags: [06 Course Intakes]
      operationId: listCoursesAdmin
      summary: List courses for administration
      x-access: ADMIN
      parameters:
        - $ref: "#/components/parameters/ServiceIdFilter"
        - $ref: "#/components/parameters/CategoryIdFilter"
        - name: status
          in: query
          schema: { type: string, enum: [DRAFT, OPEN_ACTIVE, CLOSED_ACTIVE, COMPLETED, CANCELLED, ARCHIVED] }
        - $ref: "#/components/parameters/CourseLevelFilter"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          $ref: "#/components/responses/Success"
    post:
      tags: [06 Course Intakes]
      operationId: createCourse
      summary: Create a course intake draft
      x-access: ADMIN
      description: The first intake defines immutable public content. Certificate policy comes from the course group. Later intakes copy a selected existing sibling and its active Ready curriculum into independent course-session rows; learner history is never copied.
      requestBody:
        $ref: "#/components/requestBodies/CreateCourse"
      responses:
        "201":
          $ref: "#/components/responses/Created"
  /api/v1/courses/{courseId}:
    get:
      tags: [06 Course Intakes]
      operationId: getCourseAdmin
      summary: Get course for administration
      x-access: ADMIN
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          $ref: "#/components/responses/Success"
    patch:
      tags: [06 Course Intakes]
      operationId: updateCourse
      summary: Update intake dates, timezone, or capacity
      x-access: ADMIN
      description: Certificate policy is not editable. It was fixed when the course was created.
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        $ref: "#/components/requestBodies/UpdateCourse"
      responses:
        "200":
          $ref: "#/components/responses/Success"
    delete:
      tags: [06 Course Intakes]
      operationId: deleteCoursePermanently
      summary: "[DESTRUCTIVE] Permanently delete archived course"
      x-access: SUPER_ADMIN
      description: Super-admin only. Deletes only an archived intake without curriculum or learner history.
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/CatalogDeletionBlocked"
  /api/v1/courses/{courseId}/deletion-impact:
    get:
      tags: [06 Course Intakes]
      operationId: getCourseDeletionImpact
      summary: Preview whether permanent course deletion is safe
      x-access: SUPER_ADMIN
      description: Super-admin only. Aggregates protected curriculum and enrollment history without changing data.
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          $ref: "#/components/responses/Success"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: Set automatically by Login or Verify email OTP.
  responses:
    Success:
      description: Successful API response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
    Created:
      description: Resource created successfully
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
    CatalogDeletionBlocked:
      description: Permanent deletion would destroy operational delivery or protected learner history.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CatalogDeletionBlockedEnvelope"
  parameters:
...
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema: { type: string, format: uuid }
      example: "{{categoryId}}"
    CourseId:
      name: courseId
      in: path
      required: true
      schema: { type: string, format: uuid }
      example: "{{courseId}}"
...
    ServiceIdFilter:
      name: serviceId
      in: query
      schema: { type: string, format: uuid }
      example: "{{learningServiceId}}"
...
    CategoryIdFilter:
      name: categoryId
      in: query
      schema: { type: string, format: uuid }
      example: "{{categoryId}}"
    CatalogStatusFilter:
      name: status
      in: query
      schema:
        $ref: "#/components/schemas/CatalogStatus"
      example: DRAFT
    CourseLevelFilter:
      name: level
      in: query
      schema:
        $ref: "#/components/schemas/CourseLevel"
      example: BEGINNER
...
    SearchQuery:
      name: q
      in: query
      schema: { type: string }
      example: ai
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  requestBodies:
...
    CreateCategory:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CreateCategoryInput"
          example:
            serviceId: "{{learningServiceId}}"
            slug: api-testing
            title: API Testing
            description: Practical API testing courses for software engineers.
            audienceLabel: Beginner friendly
            visualKey: code2
            badgeLabel: New
            sortOrder: 0
    UpdateCategory:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UpdateCategoryInput"
          example:
            title: API Testing and Quality
            description: Practical API testing and quality courses for software engineers.
            sortOrder: 1
...
    CreateCourse:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CreateCourseInput"
          example:
            courseGroupId: "{{courseGroupId}}"
            intakeKey: 2026-B1
            startDate: "2026-09-01T00:00:00.000Z"
            expectedEndDate: "2026-12-01T00:00:00.000Z"
            timezone: Asia/Colombo
            capacity: 50
            summary: Learn practical AI and machine-learning foundations.
            description: A practical guided program for engineers beginning applied AI and machine learning.
            level: BEGINNER
            durationValue: 4
            durationUnit: WEEK
            price: 0
            highlights: [Requests, Environments, Tests]
            skills: [Postman, HTTP, API Testing]
            prerequisites: [Basic web knowledge]
            thumbnailUrl: null
            sortOrder: 0
    UpdateCourse:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UpdateCourseInput"
          example:
            capacity: 60
            expectedEndDate: "2026-12-15T00:00:00.000Z"
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success: { type: boolean, example: true }
        data: {}
        message: { type: string, example: Success }
...
    CatalogDeletionBlockedEnvelope:
      type: object
      required: [success, error, code, details]
      properties:
        success: { type: boolean, example: false }
        error:
          type: string
          example: Permanent deletion is blocked because this catalog item contains curriculum or learner history.
        code:
          type: string
          enum: [CATALOG_DELETION_BLOCKED]
        details:
          $ref: "#/components/schemas/CatalogDeletionImpact"
    CatalogDeletionImpact:
      type: object
      required: [resourceType, resourceId, resourceStatus, deletable]
      properties:
        resourceType: { type: string, enum: [LEARNING_SERVICE, CATEGORY, COURSE_GROUP, COURSE] }
        resourceId: { type: string, format: uuid }
        resourceStatus: { type: string }
        deletable: { type: boolean }
        courses: { type: integer, minimum: 0 }
        courseGroups: { type: integer, minimum: 0 }
        history: { type: integer, minimum: 0 }
        curriculumLinks: { type: integer, minimum: 0 }
        enrollments: { type: integer, minimum: 0 }
        projects: { type: integer, minimum: 0 }
...
    CatalogStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ARCHIVED]
    CourseLevel:
      type: string
      enum: [OPEN, FOUNDATION, BEGINNER, INTERMEDIATE, ADVANCED]
...
    DurationUnit:
      type: string
      enum: [SESSION, DAY, WEEK, MONTH]
...
    CreateCategoryInput:
      type: object
      required: [serviceId, slug, title, description, visualKey]
      properties:
        serviceId: { type: string, format: uuid }
        slug: { type: string, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$" }
        title: { type: string, minLength: 2, maxLength: 120 }
        description: { type: string, minLength: 10, maxLength: 600 }
        audienceLabel: { type: string, nullable: true }
        visualKey:
          type: string
          enum: [sparkles, cpu, code2, workflow, calculator, atom, bar-chart3, terminal, git-branch, globe, languages]
        badgeLabel: { type: string, nullable: true }
        sortOrder: { type: integer, minimum: 0 }
    UpdateCategoryInput:
      type: object
      minProperties: 1
      properties:
        slug: { type: string, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$" }
        title: { type: string, minLength: 2, maxLength: 120 }
        description: { type: string, minLength: 10, maxLength: 600 }
        audienceLabel: { type: string, nullable: true }
        visualKey:
          type: string
          enum: [sparkles, cpu, code2, workflow, calculator, atom, bar-chart3, terminal, git-branch, globe, languages]
        badgeLabel: { type: string, nullable: true }
        sortOrder: { type: integer, minimum: 0 }
...
    CreateCourseInput:
      type: object
      required: [courseGroupId, intakeKey]
      properties:
        courseGroupId: { type: string, format: uuid }
        sourceCourseId: { type: string, format: uuid, description: Required after the first intake in a group. }
        intakeKey: { type: string, pattern: "^[A-Z0-9]+(?:-[A-Z0-9]+)*$" }
        startDate: { type: string, format: date-time, nullable: true }
        expectedEndDate: { type: string, format: date-time, nullable: true }
        timezone: { type: string, default: Asia/Colombo }
        capacity: { type: integer, minimum: 1, maximum: 100000, nullable: true }
        summary: { type: string, minLength: 10, maxLength: 300 }
        description: { type: string, minLength: 20, maxLength: 5000 }
        level:
          $ref: "#/components/schemas/CourseLevel"
        durationValue: { type: integer, minimum: 1, maximum: 999, nullable: true }
        durationUnit:
          allOf:
            - $ref: "#/components/schemas/DurationUnit"
          nullable: true
        price: { type: number, minimum: 0, maximum: 99999999 }
        highlights:
          type: array
          maxItems: 30
          items: { type: string }
        skills:
          type: array
          maxItems: 30
          items: { type: string }
        prerequisites:
          type: array
          maxItems: 30
          items: { type: string }
        thumbnailUrl: { type: string, format: uri, nullable: true }
        sortOrder: { type: integer, minimum: 0 }
    UpdateCourseInput:
      type: object
      minProperties: 1
      properties:
        startDate: { type: string, format: date-time, nullable: true }
        expectedEndDate: { type: string, format: date-time, nullable: true }
        timezone: { type: string }
        capacity: { type: integer, minimum: 1, maximum: 100000, nullable: true }
...